android: Refactor paths in Android.mk
Using source paths relative to android/ directory created mess in
Android's obj directory since object files are created also relative
to bluetoothd_intermediates (thus can be created outside it).

To avoid this we set LOCAL_PATH to absolute path (in terms of Android
build root) of directory in which bluez/ is located and change and
update source paths accordingly.

This also allows us to build in future glib and sbc easily using
Android.mk in BlueZ.