android: Cancel all pending confirm name before stop discovery
If user wants to cancel discovery we should remove all the confirm name
requests from mgmt queues.

It is in order to make sure that stop discovery have a free way to
reach kernel.

This improves scenario when there is a big inflow of device found events and
mgmt queues become full of confirm name requests. In such case stop discovery
might stack in the queue.