core: Ignore ATT errors when considering Pair() success
Occasionally there may be errors in completing the service discovery
process over ATT when pairing with an LE device. Since we will any way
pick up from where we left off the next time we get connected it should
be safe to return success for the Pair() method call as long as the
actual pairing part was successful.