core/profile: Fix regression when registering for client only
This got introduced with PCE that register a record for PBAP client, but
it end up causing a regression where all profiles that have records would
be probed as if server was enabled as well.

Note that is was possible to register PBAP client with PCE record, in
that case Role option should be omitted so both server and client role
are enabled by default which does not affect PBAP server component since
it uses a different UUID, in fact the correct form to add PCE record is
probably to register as server only.