tools/mgmt-tester: Fix "Pair Device - SSP Confirm Reject 1"
With changes introduced in kernel by 6fd6b915 we no longer require MITM
when being pairing initiator (i.e. we indicate it to remote but do not
force it when remote does not require it) thus kernel will auto-accept
pairing when remote also does not require MITM. As a result it's not
possible for tester to reject user confirmation since there's no
request from kernel.

This patch sets acceptor authentication requirements to include MITM
so we'll have user confirmation request and tester can reject it.