doc/obex-api: Add API for PBAP 1.2 features
These changes add support for PBAP 1.2 features, the idea is that upon
connecting the stack will auto retrieve the DatabaseIdentifier which will
very likely be used by applications to detect if anything should be done.

In case DatabaseIdentifier is present the user application can then
request the folder version via GetVersion.

Two new filter are added, Operator which should be used along with FilterBy
and ResetNewMissedCalls the later can only be used in mch and cch folders.