core: Bind connectable mode to discoverable for kernel connection control
When the kernel supports the BR/EDR whitelist we don't need to always be
connectable in order to receive incoming connections. Therefore, bind
the connectable mode to the discoverable mode whenever kernel connection
control is supported.

To avoid unnecessary mgmt and HCI commands disabling discoverable is
simply done by disabling connectable as it has the effect of disabling
both settings.