android/tester-ng: Make action calls non recurent
All action steps steps were called recurently if not interleaved with
callbacks steps in test case definition. This could block any potential
callback that was not expected to come, as mainloop was stuck. This could
raise potential stability issues and timeouts when working with file
descriptors and io channels handled in the main loop.