android/bluetooth: Fix retriggering the discovery
Previously le device found callback was set every time we triggered the scan
from gatt. But it does no longer work like that. This callback is now set
when gatt profile is registered and should not be used for checking if we
should be retriggering the scan. Expected discovery flag should be
used directly instead.