android/bluetooth: Fix handling paring related events
As it turns out it is possible that kernel send to user space mgmt
events related to pairing before device connected event.
In such case BfA was dropping such event because he does not know
device.

With this patch, if device is not in the cache, BfA will create it.

Issue triggered by test: "Bluetooth Accept Bond - Just Works - Success"