android/gatt: Fix adding duplicate connection for server apps
This replaces the usage of unsafe function for adding app connections
with the one that is almost the same but first it checks if no
duplicate entry is added to the list.

This fixes connection callback being sent twice for server app when it
initiates the connection, and removes duplicated code.