android/hal-ipc-api: Documentation update and cleanup
This patch clarifies that 0x80 opcode is reserved and not used. This is
in case we need to introduce confirmations for notifications.

Error response definition is moved into common place and all HALs are
simply referring to it. This is to keep valid status values in single
place. Valis status is also updated with 'Authentication Rejected'
error added in Android 5.0 API.