shared/att: Add API for sending ATT error response
This patch adds bt_att_send_error_rsp which sends an ATT protocol error
response PDU over the underlying link based on the parameters that are
passed to it. This is a useful function since it reduces a lot of
repeated code in shared/gatt-server that does the same thing.

The "error" parameter is expressed as a signed integer (in line with the
new shared/gatt-db APIs). bt_att_send_error_rsp correctly maps this
(including possible UNIX errnos) to ATT protocol error codes. In
addition, struct bt_att_pdu_error_rsp is introduced to represent a
packed error response PDU which can be sent directly over the wire once
properly encoded in network order.