shared/gatt-client: Fix memory bug in service_changed_complete
This patch fixes a bug that got accidentally introduced in a previous
memory leak fix, where the temporary discovery_op structure held on to
the newly discovered service pointers after passing their ownership to
the client and incorrectly free'd their memory. This was correctly fixed
in init_complete but missed in service_changed_complete.