gattrib: Fix not setting att's mtu size in g_attrib_new()
We were setting only the buffer in gattrib but att was left with the
default value of 23 (LE transport layer defaults). For BREDR transport
layer this may result in using smaller MTU than the minimum allowed
(48).

This was affecting cases when MTU Exchange was not supported by the
remote device and defaults were used.