android/gatt: Fix embedded services not responsing to some requests
For application attributes this is triggered when application sends
response but for embedded ones we need to trigger this manually. As in
prepare and execute write requests handlers, we don't know if it is
embedded or application attribute so we should try sending processed
request responses if there are any.