android/tester: Refractor properties veryfying
Pre condition for properties veryfying is to have set expected
properties in step - there is no need to check it inside veryfying
function. However veryfying function should check if number of expected
or received number of properties is valid. If properties are expected
but number of them is 0, test should fail.