android/tester: Add macro improving pdu send procedure
As of now when we want to send response from remote device, we compose
pdu from handle and value, which are then copied into fixed size pdu
with predefined header. With this macro we will be able to pass up to
3 iovectors which are then send without additional copying. It also
quite generic and can be used to any other kind of data processing nad
not only for sending pdus.