emulator: Fix L2CAP & RFCOMM state changes
We should only move to do RFCOMM operations once we exist the CONFIG
state. This means not doing anything if we get a "pending" response and
moving to RFCOMM only once the config req/rsp messages are exchanged.
This patch still isn't a "complete" fix in that it only waits for a
response to our config request, but it's good enough for now and makes
the RFCOMM server tests in rfcomm-tester pass again.