shared/gatt: Allow register_notify without CCC
Most OSes allow to register characteristics without a CCC even if they
do support notification/indication, this is not according to the
specification and may not even work in practice since there is no way
to enable notifications/indications without a CCC to write to but there
has been some applications that do that anyway.

This patch fixes this behavior so that bt_gatt_client_register_notify
immediately registers the callback and returns success for
characteristics with no CCC, this does cause any air traffic therefore
it is safe to assume it won't cause any problems.