audioo/AVCTP: Set initial key repeat timeout
'AVC Panel Subunit Specification' mandates that:
To help the target to decide whether a pressed button is hold
or not, it is recommended that the controller have capability
of issuing a “release” command within 300 ms, after it receives
a response to the last “press” command. The target should not
judge the user operation to be “press and hold” within this
300ms.". The kernel's default timeout is currently set to 250ms
with 33ms between repeats which means it will issue one repeat
key before 300ms window elapses.