doc/mgmt-api: Update error codes for Pair Device Command
This patch adds two error codes for Pair device command.

Reject error code which is used when requested transport is not enabled.
E.g. User wants to do LE pair but LE support is not enabled. Similar
with BREDR

Not Supported error code which is used if controller is not capable with
requested transport