android/bluetooth: Fix enabling secure connection
Current code works fine for controllers supporting BR/EDR secure
connections.

We want to have LE Secure connection enabled also for controllers which
does not support BR/EDR secure connections. For this we need to set
secure connection in the kernel just after LE is enabled (if requested
on register)