audio/avrcp: Fix crash when acting as TG
When acting as TG the code may end up using GetCapabilities to figure
out if volume changed event is supported, this cause crashes with
remote devices that don't support TG themselves because
session->controller is not initialized.

Note that to pass AVRCP qualification when volume changed is supported
the device has to support TG as well as CT record, this is because PTS
will attempt to fetch the TG record for any test that it is sending
commands so devices where this crash happen most likely don't have
volume control feature enabled.