tools/mgmt-tester: expect 0 rp when removing all adv inst
The kernel would previously return a hard coded instance value of 0x01
even when removing multiple advertising instances. The correct behavior,
though, would be to return zero when multiple instances have been
removed. This corresponds to the semantics of the mgmt API call made in
the first place.

A fix for this problem has been introduced in the kernel. Now the
corresponding test is updated to reflect that logic.